# Improve deferred write encryption