- Uses a per-volume pool, instead of a global pool (#1285)
- Make encryption infallible; because the only failure conditions are for sizes that we can't hit (and this makes the next step easier)
- Don't defer encryption for small writes, using the same logic as reads