@mkeeter 2024-05-14
Doing a quick in-VM perf check with 8 worker threads (6eb75d45), this seems basically fine

## `main`
```
4M WRITE: bw=809MiB/s (848MB/s), 809MiB/s-809MiB/s (848MB/s-848MB/s), io=47.6GiB (51.1GB), run=60249-60249msec
4K WRITE: bw=10.5MiB/s (11.0MB/s), 10.5MiB/s-10.5MiB/s (11.0MB/s-11.0MB/s), io=631MiB (661MB), run=60012-60012msec
1M WRITE: bw=735MiB/s (771MB/s), 735MiB/s-735MiB/s (771MB/s-771MB/s), io=43.1GiB (46.3GB), run=60060-60060msec
4M WRITE: bw=825MiB/s (865MB/s), 825MiB/s-825MiB/s (865MB/s-865MB/s), io=48.4GiB (51.9GB), run=60057-60057msec
4K READ: bw=19.8MiB/s (20.8MB/s), 19.8MiB/s-19.8MiB/s (20.8MB/s-20.8MB/s), io=1190MiB (1248MB), run=60005-60005msec
1M READ: bw=929MiB/s (974MB/s), 929MiB/s-929MiB/s (974MB/s-974MB/s), io=54.4GiB (58.4GB), run=60028-60028msec
4M READ: bw=1019MiB/s (1068MB/s), 1019MiB/s-1019MiB/s (1068MB/s-1068MB/s), io=59.8GiB (64.2GB), run=60095-60095msec
```

## This branch
```
4M WRITE: bw=803MiB/s (842MB/s), 803MiB/s-803MiB/s (842MB/s-842MB/s), io=47.1GiB (50.6GB), run=60094-60094msec
4K WRITE: bw=13.4MiB/s (14.1MB/s), 13.4MiB/s-13.4MiB/s (14.1MB/s-14.1MB/s), io=806MiB (846MB), run=60013-60013msec
1M WRITE: bw=795MiB/s (833MB/s), 795MiB/s-795MiB/s (833MB/s-833MB/s), io=46.6GiB (50.0GB), run=60026-60026msec
4M WRITE: bw=814MiB/s (853MB/s), 814MiB/s-814MiB/s (853MB/s-853MB/s), io=47.7GiB (51.3GB), run=60076-60076msec
4K READ: bw=19.8MiB/s (20.8MB/s), 19.8MiB/s-19.8MiB/s (20.8MB/s-20.8MB/s), io=1189MiB (1246MB), run=60006-60006msec
1M READ: bw=943MiB/s (988MB/s), 943MiB/s-943MiB/s (988MB/s-988MB/s), io=55.3GiB (59.3GB), run=60027-60027msec
4M READ: bw=1033MiB/s (1084MB/s), 1033MiB/s-1033MiB/s (1084MB/s-1084MB/s), io=60.7GiB (65.1GB), run=60100-60100msec
```

This is all well within the normal range of variation.